fix: fix dependabot update (#565)
