+++ backend/workers/llm-finetuning-node/datasets/requirements.txt
@@ -6,10 +6,11 @@ requests==2.33.0
 pydantic==2.10.6
 
 # Vector dependencies
-langchain==0.3.30
-langchain-chroma==0.2.5
-langchain-community===0.3.27
-chromadb==1.0.20
+langchain==1.3.11
+langchain-chroma==1.1.0
+langchain-community===0.4.2
+langchain-text-splitters==1.1.2
+chromadb==1.5.9
 pysqlite3-binary==0.5.4
 FlagEmbedding==1.3.5 
 pypdf>=6.7.4
