--- backend/workers/llm-finetuning-node/telemetry/entrypoint.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-# Copyright (C) 2025 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0 
-
-
-echo "Starting app ..."
-exec "$@"
