--- backend/workers/llm-finetuning-node/telemetry/requirements.txt
@@ -1,11 +1,0 @@
-# Celery dependencies
-celery==5.4.0
-celery[redis]==5.4.0
-eventlet==0.40.3
-python-dotenv==1.2.2
-requests==2.33.0
-
-# System info dependencies
-py-cpuinfo==9.0.0
-psutil==6.0.0
-pydantic==2.10.6
