+++ frontend/package.json
@@ -13,9 +13,8 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@ai-sdk/openai": "^1.3.18",
-    "@ai-sdk/openai-compatible": "^1.0.29",
-    "@ai-sdk/react": "^2.0.115",
+    "@ai-sdk/openai-compatible": "^2.0.51",
+    "@ai-sdk/react": "^3.0.210",
     "@hookform/resolvers": "^4.1.3",
     "@radix-ui/react-accordion": "^1.2.3",
     "@radix-ui/react-alert-dialog": "^1.1.6",
@@ -38,9 +37,8 @@
     "@tanstack/react-query": "^5.74.3",
     "@tanstack/react-query-devtools": "^5.74.3",
     "@types/recharts": "^1.8.29",
-    "ai": "^5.0.113",
+    "ai": "^6.0.208",
     "axios": "^1.13.5",
-    "celery-ts": "^1.2.3",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.477.0",
